+++ CHANGELOG.md
@@ -10,6 +10,37 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Session audit metrics
 
+## [1.43.3] - 2024-07-31
+
+### Changed
+- Changed the token validation directive of some operations
+
+## [1.43.2] - 2024-07-29
+
+### Added
+- Add enforcement of new validation for admin and api tokens
+- Add more details to admin and api token validation metric
+
+## [1.43.1] - 2024-07-24
+
+### Changed
+- Changed the token validation directive of some operations
+
+## [1.43.0] - 2024-07-23
+
+### Added
+- Add admin validation directive
+
+## [1.42.0] - 2024-07-17
+
+### Fixed
+- Get tokens from headers when necessary
+
+## [1.41.1] - 2024-07-15
+
+### Added
+- Add validation metrics for admin and api tokens
+
 ## [1.41.0] - 2024-07-01
 
 ### Added
@@ -26,6 +57,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Check user is part of buyer org instead of "active" on checkUserAccess directive
 
 ## [1.40.5] - 2024-05-22
+
 ### Changed
 - Improved metrics and logging for checkUserAccess and checkAdminAccess directives